get proxy deployments working
